+++ src/Musurp/Neo/Cypher/Tests/Unit/Builder/QueryBuilderTest.php
@@ -12,6 +12,8 @@ declare(strict_types=1);
 namespace Musurp\Neo\Cypher\Tests\Unit\Builder;
 
 use Musurp\Neo\Cypher\Builder\QueryBuilder;
+use Musurp\Neo\Cypher\Component\Clause\MatchClause;
+use Musurp\Neo\Cypher\Component\Clause\OptionalMatchClause;
 
 use PHPUnit\Framework\TestCase;
 
@@ -64,6 +66,26 @@ CYPHER;
      *
      * @covers \Musurp\Neo\Cypher\Builder\QueryBuilder
      */
+    public function createQueryBuilderWithSingleOptionalMatch(): void
+    {
+        $builder = new QueryBuilder();
+        $builder->match($builder::path()->create(null, [], []), true);
+
+        $cypher = <<<CYPHER
+OPTIONAL MATCH ()
+CYPHER;
+
+        self::assertEquals($cypher, $builder->build());
+    }
+
+    /**
+     * @test
+     *
+     * @group unit
+     * @group component
+     *
+     * @covers \Musurp\Neo\Cypher\Builder\QueryBuilder
+     */
     public function createQueryBuilderWithMultipleMatches(): void
     {
         $builder = new QueryBuilder();
@@ -73,6 +95,28 @@ CYPHER;
         $cypher = <<<CYPHER
 MATCH ()
 MATCH (:ONE)
+CYPHER;
+
+        self::assertEquals($cypher, $builder->build());
+    }
+
+    /**
+     * @test
+     *
+     * @group unit
+     * @group component
+     *
+     * @covers \Musurp\Neo\Cypher\Builder\QueryBuilder
+     */
+    public function createQueryBuilderWithMultipleOptionalMatches(): void
+    {
+        $builder = new QueryBuilder();
+        $builder->match($builder::path()->create(null, [], []), true);
+        $builder->match($builder::path()->create(null, ['ONE'], []), true);
+
+        $cypher = <<<CYPHER
+OPTIONAL MATCH ()
+OPTIONAL MATCH (:ONE)
 CYPHER;
 
         self::assertEquals($cypher, $builder->build());
@@ -102,6 +146,56 @@ MATCH ()
 MATCH
   (:ONE),
   (var:ONE {foo: 'bar'})
+CYPHER;
+
+        self::assertEquals($cypher, $builder->build());
+    }
+
+    /**
+     * @test
+     *
+     * @group unit
+     * @group component
+     *
+     * @covers \Musurp\Neo\Cypher\Builder\QueryBuilder
+     */
+    public function createQueryBuilderWithMatchInstance(): void
+    {
+        $builder = new QueryBuilder();
+        $builder->match(
+            new MatchClause([
+                $builder::path()->node(null, [], []),
+            ])
+        );
+
+        $cypher = <<<CYPHER
+MATCH ()
+CYPHER;
+
+        self::assertEquals($cypher, $builder->build());
+    }
+
+    /**
+     * @test
+     *
+     * @group unit
+     * @group component
+     *
+     * @covers \Musurp\Neo\Cypher\Builder\QueryBuilder
+     */
+    public function createQueryBuilderWithOptionalMatchInstance(): void
+    {
+        $builder = new QueryBuilder();
+        $builder->match(
+            new OptionalMatchClause(
+                new MatchClause([
+                    $builder::path()->node(null, [], []),
+                ])
+            )
+        );
+
+        $cypher = <<<CYPHER
+OPTIONAL MATCH ()
 CYPHER;
 
         self::assertEquals($cypher, $builder->build());